move api helpers into utils
