+++ frontend/src/components/Footer.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { getLogoutUrl } from '../api'
+import { getLogoutUrl } from '../utils/api'
 
 /**
  * The footer of every page.
